requirements.txt
langchain-community
langchain-groq
langchain
langchain-core
python-dotenv
streamlit
uvicorn
fastapi
pydantic
tavily-python
